+++ Tests/GRDBTests/PersistableTests.swift
@@ -146,7 +146,7 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM persons")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
         }
     }
 
@@ -158,7 +158,7 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM persons")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
         }
     }
 
@@ -173,8 +173,8 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM persons")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "id") as Int64, person.id!)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
+            XCTAssertEqual(rows[0]["id"] as Int64, person.id!)
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
         }
     }
 
@@ -195,10 +195,10 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Craig")
-            XCTAssertEqual(rows[1].value(named: "id") as Int64, person2.id!)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
+            XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+            XCTAssertEqual(rows[0]["name"] as String, "Craig")
+            XCTAssertEqual(rows[1]["id"] as Int64, person2.id!)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
         }
     }
 
@@ -217,12 +217,12 @@ class PersistableTests: GRDBTestCase {
                 
                 let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
                 XCTAssertEqual(rows.count, 2)
-                XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-                XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
-                XCTAssertEqual(rows[0].value(named: "age") as Int, 24)
-                XCTAssertEqual(rows[1].value(named: "id") as Int64, person2.id!)
-                XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-                XCTAssertEqual(rows[1].value(named: "age") as Int, 36)
+                XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+                XCTAssertEqual(rows[0]["name"] as String, "Arthur")
+                XCTAssertEqual(rows[0]["age"] as Int, 24)
+                XCTAssertEqual(rows[1]["id"] as Int64, person2.id!)
+                XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+                XCTAssertEqual(rows[1]["age"] as Int, 36)
             }
             
             do {
@@ -233,12 +233,12 @@ class PersistableTests: GRDBTestCase {
                 
                 let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
                 XCTAssertEqual(rows.count, 2)
-                XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-                XCTAssertEqual(rows[0].value(named: "name") as String, "Craig")
-                XCTAssertEqual(rows[0].value(named: "age") as Int, 24)
-                XCTAssertEqual(rows[1].value(named: "id") as Int64, person2.id!)
-                XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-                XCTAssertEqual(rows[1].value(named: "age") as Int, 36)
+                XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+                XCTAssertEqual(rows[0]["name"] as String, "Craig")
+                XCTAssertEqual(rows[0]["age"] as Int, 24)
+                XCTAssertEqual(rows[1]["id"] as Int64, person2.id!)
+                XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+                XCTAssertEqual(rows[1]["age"] as Int, 36)
             }
             
             do {
@@ -248,12 +248,12 @@ class PersistableTests: GRDBTestCase {
                 
                 let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
                 XCTAssertEqual(rows.count, 2)
-                XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-                XCTAssertEqual(rows[0].value(named: "name") as String, "Craig")
-                XCTAssertEqual(rows[0].value(named: "age") as Int, 25)
-                XCTAssertEqual(rows[1].value(named: "id") as Int64, person2.id!)
-                XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-                XCTAssertEqual(rows[1].value(named: "age") as Int, 36)
+                XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+                XCTAssertEqual(rows[0]["name"] as String, "Craig")
+                XCTAssertEqual(rows[0]["age"] as Int, 25)
+                XCTAssertEqual(rows[1]["id"] as Int64, person2.id!)
+                XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+                XCTAssertEqual(rows[1]["age"] as Int, 36)
             }
         }
     }
@@ -266,8 +266,8 @@ class PersistableTests: GRDBTestCase {
             
             var rows = try Row.fetchAll(db, "SELECT * FROM persons")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
+            XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
             
             let person2 = PersistablePersonClass(id: nil, name: "Barbara", age: 39)
             try person2.save(db)
@@ -277,20 +277,20 @@ class PersistableTests: GRDBTestCase {
             
             rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Craig")
-            XCTAssertEqual(rows[1].value(named: "id") as Int64, person2.id!)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
+            XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+            XCTAssertEqual(rows[0]["name"] as String, "Craig")
+            XCTAssertEqual(rows[1]["id"] as Int64, person2.id!)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
             
             try person1.delete(db)
             try person1.save(db)
             
             rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "id") as Int64, person1.id!)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Craig")
-            XCTAssertEqual(rows[1].value(named: "id") as Int64, person2.id!)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
+            XCTAssertEqual(rows[0]["id"] as Int64, person1.id!)
+            XCTAssertEqual(rows[0]["name"] as String, "Craig")
+            XCTAssertEqual(rows[1]["id"] as Int64, person2.id!)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
         }
     }
 
@@ -309,8 +309,8 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY id")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "id") as Int64, person2.id!)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Barbara")
+            XCTAssertEqual(rows[0]["id"] as Int64, person2.id!)
+            XCTAssertEqual(rows[0]["name"] as String, "Barbara")
         }
     }
 
@@ -337,8 +337,8 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France")
         }
     }
 
@@ -356,10 +356,10 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
         }
     }
 
@@ -378,10 +378,10 @@ class PersistableTests: GRDBTestCase {
                 
                 let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
                 XCTAssertEqual(rows.count, 2)
-                XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-                XCTAssertEqual(rows[0].value(named: "name") as String, "France")
-                XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-                XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+                XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+                XCTAssertEqual(rows[0]["name"] as String, "France")
+                XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+                XCTAssertEqual(rows[1]["name"] as String, "United States")
             }
             
             do {
@@ -391,10 +391,10 @@ class PersistableTests: GRDBTestCase {
                 
                 let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
                 XCTAssertEqual(rows.count, 2)
-                XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-                XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-                XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-                XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+                XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+                XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+                XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+                XCTAssertEqual(rows[1]["name"] as String, "United States")
             }
         }
     }
@@ -407,8 +407,8 @@ class PersistableTests: GRDBTestCase {
             
             var rows = try Row.fetchAll(db, "SELECT * FROM countries")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France")
             
             let country2 = PersistableCountry(isoCode: "US", name: "United States")
             try country2.save(db)
@@ -418,20 +418,20 @@ class PersistableTests: GRDBTestCase {
             
             rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
             
             try country1.delete(db)
             try country1.save(db)
             
             rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
         }
     }
 
@@ -450,8 +450,8 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[0]["name"] as String, "United States")
         }
     }
 
@@ -496,8 +496,8 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France")
         }
     }
 
@@ -540,10 +540,10 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
         }
     }
 
@@ -586,10 +586,10 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
         }
     }
 
@@ -619,8 +619,8 @@ class PersistableTests: GRDBTestCase {
             
             var rows = try Row.fetchAll(db, "SELECT * FROM countries")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France")
             
             let country2 = PersistableCustomizedCountry(
                 isoCode: "US",
@@ -643,10 +643,10 @@ class PersistableTests: GRDBTestCase {
             
             rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
             
             _ = try country1.delete(db)
             try country1.save(db)
@@ -659,10 +659,10 @@ class PersistableTests: GRDBTestCase {
             
             rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "FR")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "France Métropolitaine")
-            XCTAssertEqual(rows[1].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[1].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "FR")
+            XCTAssertEqual(rows[0]["name"] as String, "France Métropolitaine")
+            XCTAssertEqual(rows[1]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[1]["name"] as String, "United States")
         }
     }
 
@@ -706,8 +706,8 @@ class PersistableTests: GRDBTestCase {
             
             let rows = try Row.fetchAll(db, "SELECT * FROM countries ORDER BY isoCode")
             XCTAssertEqual(rows.count, 1)
-            XCTAssertEqual(rows[0].value(named: "isoCode") as String, "US")
-            XCTAssertEqual(rows[0].value(named: "name") as String, "United States")
+            XCTAssertEqual(rows[0]["isoCode"] as String, "US")
+            XCTAssertEqual(rows[0]["name"] as String, "United States")
         }
     }