+++ Tests/GRDBTests/UpdateStatementTests.swift
@@ -66,10 +66,10 @@ class UpdateStatementTests : GRDBTestCase {
         try dbQueue.inDatabase { db in
             let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY name")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
-            XCTAssertEqual(rows[0].value(named: "age") as Int, 41)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-            XCTAssertTrue((rows[1].value(named: "age") as DatabaseValue).isNull)
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
+            XCTAssertEqual(rows[0]["age"] as Int, 41)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+            XCTAssertTrue((rows[1]["age"] as DatabaseValue).isNull)
         }
     }
 
@@ -94,10 +94,10 @@ class UpdateStatementTests : GRDBTestCase {
         try dbQueue.inDatabase { db in
             let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY name")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
-            XCTAssertEqual(rows[0].value(named: "age") as Int, 41)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-            XCTAssertTrue((rows[1].value(named: "age") as DatabaseValue).isNull)
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
+            XCTAssertEqual(rows[0]["age"] as Int, 41)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+            XCTAssertTrue((rows[1]["age"] as DatabaseValue).isNull)
         }
     }
 
@@ -121,10 +121,10 @@ class UpdateStatementTests : GRDBTestCase {
         try dbQueue.inDatabase { db in
             let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY name")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
-            XCTAssertEqual(rows[0].value(named: "age") as Int, 41)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-            XCTAssertTrue((rows[1].value(named: "age") as DatabaseValue).isNull)
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
+            XCTAssertEqual(rows[0]["age"] as Int, 41)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+            XCTAssertTrue((rows[1]["age"] as DatabaseValue).isNull)
         }
     }
 
@@ -149,10 +149,10 @@ class UpdateStatementTests : GRDBTestCase {
         try dbQueue.inDatabase { db in
             let rows = try Row.fetchAll(db, "SELECT * FROM persons ORDER BY name")
             XCTAssertEqual(rows.count, 2)
-            XCTAssertEqual(rows[0].value(named: "name") as String, "Arthur")
-            XCTAssertEqual(rows[0].value(named: "age") as Int, 41)
-            XCTAssertEqual(rows[1].value(named: "name") as String, "Barbara")
-            XCTAssertTrue((rows[1].value(named: "age") as DatabaseValue).isNull)
+            XCTAssertEqual(rows[0]["name"] as String, "Arthur")
+            XCTAssertEqual(rows[0]["age"] as Int, 41)
+            XCTAssertEqual(rows[1]["name"] as String, "Barbara")
+            XCTAssertTrue((rows[1]["age"] as DatabaseValue).isNull)
         }
     }